requirements.txt
typer[all]==0.20.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.1
python-dotenv>=1.0.0
rich>=14.2.0
google-api-core>=2.11.0
chromadb>=0.3.21
langchain-community>=0.0.1
--only-binary :all: pyarrow==11.0.0
